refactoring and remove unused imports
